+++ product_code/user_service/api/src/main/scala/de/upb/cs/uc4/user/model/user/Lecturer.scala
@@ -36,7 +36,8 @@ case class Lecturer(
   override def trim: Lecturer =
     super.trim.asInstanceOf[Lecturer].copy(freeText = freeText.trim, researchArea = researchArea.trim)
 
-  override def toPublic: Lecturer = super.toPublic.asInstanceOf[Lecturer]
+  override def toPublic: Lecturer =
+    Lecturer(this.username, this.role, Address.empty, this.firstName, this.lastName, this.email, this.phoneNumber, "", this.freeText, this.researchArea)
 
   override def clean: Lecturer = super.clean.asInstanceOf[Lecturer]